refactor pagination
