+++ src/css/style.css
@@ -443,7 +443,7 @@ color: #ccc;
   display: flex;
   flex-direction: column;
   gap: .3em;
-
+  padding: 1em;
 }
 
 .flight-time {
