+++ src/js/flight-List.js
@@ -1,5 +1,6 @@
 import {
   loadDynamicContent,
+  setBackgroundImage,
 } from './utilities.mjs'
 
 import {
@@ -15,169 +16,140 @@ import {
   flightCardTemplate,
 } from './flightCard.mjs'
 
+import {
+  renderPagination,
+} from './paginationHelper.mjs';
+
 loadDynamicContent ();
 toggleReturnDate();
 let currencyData = [];
+let fetchedFlights = [];
 const flightsContainer = document.getElementById('flightsContainer');
-
-const h1Tag = document.querySelector('h1');
-const paginationContainer = document.createElement('div');
-paginationContainer.className = 'pagination';
-h1Tag.insertAdjacentElement('afterend', paginationContainer);
 
 // Initialization (Assuming the page starts with the first page)
 let currentPage = parseInt(sessionStorage.getItem('currentPage'), 10) || 1;
 console.log('Current Page on load:', currentPage);
 document.getElementById('searchFlights').click();
-
-
-// Render Pagination Function
-function renderPagination(page, totalPg) {
-  paginationContainer.innerHTML = '';
-
-  let pages = [1];
-  for (let i = page - 2; i <= page + 2; i++) {
-      if (i > 1 && i < totalPg) {
-          pages.push(i);
-      }
-  }
-  if (totalPg !== 1) {
-      pages.push(totalPg);
-  }
-
-  if (page > 1) {
-      let prevLink = document.createElement('a');
-      prevLink.href = '#';
-      prevLink.innerText = 'Previous';
-      prevLink.dataset.page = page - 1;
-      paginationContainer.appendChild(prevLink);
-  }
-
-  let lastPageRendered = 0;
-  for (let p of pages) {
-      if (lastPageRendered < p - 1) {
-          let ellipsis = document.createElement('span');
-          ellipsis.innerText = '...';
-          paginationContainer.appendChild(ellipsis);
-      }
-
-      let pageLink = document.createElement('a');
-      pageLink.href = '#';
-      pageLink.innerText = p;
-      pageLink.dataset.page = p;
-
-      if (p === page) {
-          pageLink.classList.add('current');
-      }
-
-      paginationContainer.appendChild(pageLink);
-      lastPageRendered = p;
-  }
-
-  if (page < totalPg) {
-      let nextLink = document.createElement('a');
-      nextLink.href = '#';
-      nextLink.innerText = 'Next';
-      nextLink.dataset.page = page + 1;
-      paginationContainer.appendChild(nextLink);
-  }
-  // Store pagination data in sessionStorage
-  sessionStorage.setItem('currentPage', page.toString());
-  sessionStorage.setItem('totalPages', totalPg.toString());
-}
-
-// Event listener for pagination links
-paginationContainer.addEventListener('click', function(event) {
-  if (event.target.tagName === 'A') {
-      event.preventDefault();
-      const pageNumber = parseInt(event.target.dataset.page, 10);
-      currentPage = pageNumber;  // Update the currentPage variable
-      sessionStorage.setItem('currentPage', currentPage);
-      // Trigger the searchFlights click event to fetch data for the new page
-      document.getElementById('searchFlights').click();
-  }
-});
+let paginationContainer;
 
 // Retrieving and Rendering Pagination on Page Load
 document.addEventListener('DOMContentLoaded', function() {
+  const h1Tag = document.querySelector('h1');
+  paginationContainer = document.createElement('div');
+  paginationContainer.className = 'pagination';
+  h1Tag.insertAdjacentElement('afterend', paginationContainer);
+
   const storedCurrentPage = parseInt(sessionStorage.getItem('currentPage'), 10);
   const storedTotalPages = parseInt(sessionStorage.getItem('totalPages'), 10);
 
   if (storedCurrentPage && storedTotalPages) {
-      renderPagination(storedCurrentPage, storedTotalPages);
+      renderPagination(storedCurrentPage, storedTotalPages, paginationContainer);
   }
+  // Event listener for pagination links
+  paginationContainer.addEventListener('click', function(event) {
+    if (event.target.tagName === 'A') {
+        event.preventDefault();
+        const pageNumber = parseInt(event.target.dataset.page, 10);
+        currentPage = pageNumber;  // Update the currentPage variable
+        sessionStorage.setItem('currentPage', currentPage);
+        // Trigger the searchFlights click event to fetch data for the new page
+        document.getElementById('searchFlights').click();
+    }
+  });
 });
 
-// Capture form inputs, fetch new flight data, and store it in session storage
-document.getElementById('searchFlights').addEventListener('click', async function(event) {
+// Helper function to get input values
+function getInputValue(id) {
+  return document.getElementById(id).value;
+}
+
+// Helper function to construct the API URL
+function constructApiURL(searchParameters) {
+  const {
+    sourceCode,
+    destinationCode,
+    departureDate,
+    returnDate,
+    itineraryType,
+    sortFlights,
+    numAdults,
+    numSeniors,
+    classOfService,
+    nonstop,
+    currencyCode,
+  } = searchParameters;
+
+  const returnDateParam = itineraryType === 'ROUND_TRIP' && returnDate ? `&returnDate=${encodeURIComponent(returnDate)}` : '';
+  return `https://${API_HOST}/api/v1/flights/searchFlights?sourceAirportCode=${sourceCode}&destinationAirportCode=${destinationCode}&date=${encodeURIComponent(departureDate)}${returnDateParam}&itineraryType=${itineraryType}&sortOrder=${sortFlights}&numAdults=${numAdults}&numSeniors=${numSeniors}&classOfService=${classOfService}&pageNumber=${currentPage.toString()}&nonstop=${nonstop}&currencyCode=${currencyCode}`;
+}
+
+// Event listener for form submission
+document.getElementById('searchFlights').addEventListener('click', async function (event) {
   event.preventDefault();  // Prevent form submission
 
   // Capture input values
-  const sourceAirport = document.getElementById('sourceAirport').value;
-  const destinationAirport = document.getElementById('destinationAirport').value;
-  const departureDate = document.getElementById('departure-date').value;
-  const returnDate = document.getElementById('return-date').value;
+  const sourceCode = getInputValue('sourceAirport');
+  const destinationCode = getInputValue('destinationAirport');
+  const departureDate = getInputValue('departure-date');
+  const returnDate = getInputValue('return-date');
   const itineraryType = document.querySelector('input[name="trip-type"]:checked').value;
-  const numAdults = document.getElementById('numAdults').value;
-  const numSeniors = document.getElementById('numSeniors').value;
-  const nonstop = document.getElementById('Nonstop').value;
-  const sortFlights = document.getElementById('sort-flights').value;
-  const classOfService = document.getElementById('cabin-class').value;
-  const currency = document.getElementById('currency').value;
+  const numAdults = getInputValue('numAdults');
+  const numSeniors = getInputValue('numSeniors');
+  const nonstop = getInputValue('Nonstop');
+  const sortFlights = getInputValue('sort-flights');
+  const classOfService = getInputValue('cabin-class');
+  const currencyCode = getInputValue('currency');
 
-  // Construct search parameters
   const searchParameters = {
-      sourceCode: sourceAirport,
-      destinationCode: destinationAirport,
-      departureDate: departureDate,
-      returnDate: (itineraryType === 'ROUND_TRIP') ? returnDate : '',
-      itineraryType: itineraryType,
-      numAdults: numAdults,
-      numSeniors: numSeniors,
-      nonstop: nonstop,
-      sortFlights: sortFlights,
-      classOfService: classOfService,
-      currencyCode: currency,
-      pageNumber: currentPage.toString()
+    sourceCode,
+    destinationCode,
+    departureDate,
+    returnDate: itineraryType === 'ROUND_TRIP' ? returnDate : '',
+    itineraryType,
+    numAdults,
+    numSeniors,
+    nonstop,
+    sortFlights,
+    classOfService,
+    currencyCode,
+    pageNumber: currentPage.toString(),
   };
 
   console.log('Search parameters:', searchParameters);
+
   // Save the search parameters in session storage
   sessionStorage.setItem('searchParameters', JSON.stringify(searchParameters));
 
-  let returnDateParam = '';
-  if (searchParameters.itineraryType === 'ROUND_TRIP' && searchParameters.returnDate) {
-      returnDateParam = `&returnDate=${encodeURIComponent(searchParameters.returnDate)}`;
-  }
   // Construct the API request URL
-  const apiURL = `https://${API_HOST}/api/v1/flights/searchFlights?sourceAirportCode=${searchParameters.sourceCode}&destinationAirportCode=${searchParameters.destinationCode}&date=${encodeURIComponent(searchParameters.departureDate)}${returnDateParam}&itineraryType=${searchParameters.itineraryType}&sortOrder=${searchParameters.sortFlights}&numAdults=${searchParameters.numAdults}&numSeniors=${searchParameters.numSeniors}&classOfService=${searchParameters.classOfService}&pageNumber=${searchParameters.pageNumber}&nonstop=${searchParameters.nonstop}&currencyCode=${searchParameters.currencyCode}`;
+  const apiURL = constructApiURL(searchParameters);
 
-  // Fetch new flight data
-  const fetchedFlights = await fetchFlights(apiURL);
-  console.log('Raw fetched flights:', fetchedFlights);
-  console.log('Total number of flights:', fetchedFlights.data.totalNumResults);
-  console.log('Number of flights in current response:', fetchedFlights.data.flights.length);
+  try {
+    // Fetch new flight data
+    // const fetchedFlights = await fetchFlights(apiURL);
+    fetchedFlights = await fetchFlights(apiURL);
+    console.log('fetchedFlights:', fetchedFlights);
+    console.log('Total number of flights:', fetchedFlights.data.totalNumResults);
+    console.log('Number of flights in the current response:', fetchedFlights.data.flights.length);
 
+    if (fetchedFlights && fetchedFlights.status === true && Array.isArray(fetchedFlights.data.flights)) {
+      const structuredFlights = structureFlightData(fetchedFlights.data.flights, currentPage);
+      sessionStorage.setItem('fetchedFlights', JSON.stringify(structuredFlights));
 
-  // Check if the API response is valid and contains flight data
-  if (fetchedFlights && fetchedFlights.status === true && Array.isArray(fetchedFlights.data.flights)) {
-    const structuredFlights = structureFlightData(fetchedFlights.data.flights);
-    sessionStorage.setItem('fetchedFlights', JSON.stringify(structuredFlights));
+      // Display the new flight data in the DOM
+      handleFlights(structuredFlights);
 
-    // Display the new flight data in the DOM
-    handleFlights(structuredFlights);
-
-    // Render pagination based on current page and total pages
-    const totalPages = Math.ceil(fetchedFlights.data.totalNumResults / fetchedFlights.data.flights.length);
-    renderPagination(currentPage, totalPages);
-
-  } else if (fetchedFlights && fetchedFlights.message) {
-    // Handle error messages from the API
-    alert(fetchedFlights.message[0].returnDate || 'An error occurred.');
-  } else {
-    console.error('Unexpected flight data format:', fetchedFlights);
+      // Render pagination based on current page and total pages
+      const totalPages = Math.ceil(fetchedFlights.data.totalNumResults / fetchedFlights.data.flights.length);
+      renderPagination(currentPage, totalPages, paginationContainer);
+    } else if (fetchedFlights && fetchedFlights.message) {
+      // Handle error messages from the API
+      alert(fetchedFlights.message[0].returnDate || 'An error occurred.');
+    } else {
+      console.error('Unexpected flight data format:', fetchedFlights);
+    }
+  } catch (error) {
+    console.error('Error fetching flight data:', error);
   }
-
 });
 
 // Container where the flight cards will be appended
@@ -276,6 +248,11 @@ document.getElementById('searchSource').addEventListener('click', function() {
   }
 });
 
+// Background Image
+document.addEventListener('DOMContentLoaded', function() {
+  setBackgroundImage('main');
+});
+
 // Load currency data and set up the event listener
 loadCurrencyData().then(data => {
   currencyData = data;
@@ -297,37 +274,4 @@ loadCurrencyData().then(data => {
           document.getElementById('currency').value = 'USD';  // Default to USD if no match is found
       }
   });
-});
-
-// Background
-document.addEventListener('DOMContentLoaded', function() {
-  // Array of background images
-  const images = [
-      '/images/backdrop1.jpg',
-      '/images/backdrop2.jpg',
-      '/images/backdrop3.jpg',
-      '/images/backdrop4.jpg',
-  ];
-
-  function setBackgroundImageBasedOnTime() {
-      const hour = new Date().getHours();
-      const mainElement = document.querySelector('main');
-
-      let imageUrl;
-
-      if (hour >= 0 && hour < 6) {
-          imageUrl = images[0];
-      } else if (hour >= 6 && hour < 12) {
-          imageUrl = images[1];
-      } else if (hour >= 12 && hour < 18) {
-          imageUrl = images[2];
-      } else {
-          imageUrl = images[3];
-      }
-
-      mainElement.style.backgroundImage = `url('${imageUrl}')`;
-  }
-
-  // Set the background image when the page loads
-  setBackgroundImageBasedOnTime();
 });
