+++ src/js/flightCard.mjs
@@ -1,5 +1,5 @@
 // Main function to construct the flight card HTML.
-export function flightCardTemplate(flight, index) {
+export function flightCardTemplate(flight) {
   // Calculating total distance of the flight.
   const totalDistance = flight.segments[0].legs.reduce((sum, leg) => sum + leg.distanceInKM, 0);
   const totalDistanceMiles = Math.round(totalDistance * 0.621371).toLocaleString(); // Convert total distance to miles.
@@ -45,26 +45,7 @@ export function flightCardTemplate(flight, index) {
               <div class="purchase-link">
                   <span><a href="${flight.purchaseLinks[0].url}" target="_blank">Book Now for ${flight.purchaseLinks[0].currency} ${flight.purchaseLinks[0].totalPricePerPassenger} | ${flight.purchaseLinks[0].providerId}</a></span>
               </div>
-                <div class="container">
-                  <label for="checkbox-${index}">
-                    <div class="label">
-                      <input type="checkbox" id="checkbox-${index}">
-                      <div class="heart">
-                        <svg viewBox="0 0 544.582 544.582">
-                          <path d="M448.069,57.839c-72.675-23.562-150.781,15.759-175.721,87.898C247.41,73.522,169.303,34.277,96.628,57.839
-                		        C23.111,81.784-16.975,160.885,6.894,234.708c22.95,70.38,235.773,258.876,263.006,258.876
-                		        c27.234,0,244.801-188.267,267.751-258.876C561.595,160.732,521.509,81.631,448.069,57.839z">
-                        </svg>
-                        <svg viewBox="0 0 544.582 544.582">
-                          <path d="M448.069,57.839c-72.675-23.562-150.781,15.759-175.721,87.898C247.41,73.522,169.303,34.277,96.628,57.839
-                		        C23.111,81.784-16.975,160.885,6.894,234.708c22.95,70.38,235.773,258.876,263.006,258.876
-                		        c27.234,0,244.801-188.267,267.751-258.876C561.595,160.732,521.509,81.631,448.069,57.839z">
-                        </svg>
-                      </div>
-                    </div>
-                  </label>
-                </div>
-              </div>
+          </div>
       </div>`;
 }
