+++ src/js/main.js
@@ -1,5 +1,6 @@
 import {
   loadDynamicContent,
+  setBackgroundImage,
 } from './utilities.mjs'
 
 import {
@@ -123,34 +124,7 @@ loadCurrencyData().then(data => {
     });
 });
 
-// Background
+// Background Image
 document.addEventListener('DOMContentLoaded', function() {
-  const images = [
-      '/images/backdrop1.jpg',
-      '/images/backdrop2.jpg',
-      '/images/backdrop3.jpg',
-      '/images/backdrop4.jpg',
-  ];
-
-  function setBackgroundImageBasedOnTime() {
-      const hour = new Date().getHours();
-      const mainElement = document.querySelector('main');
-
-      let imageUrl;
-
-      if (hour >= 0 && hour < 6) {
-          imageUrl = images[0];
-      } else if (hour >= 6 && hour < 12) {
-          imageUrl = images[1];
-      } else if (hour >= 12 && hour < 18) {
-          imageUrl = images[2];
-      } else {
-          imageUrl = images[3];
-      }
-
-      mainElement.style.backgroundImage = `url('${imageUrl}')`;
-  }
-
-  // Set the background image when the page loads
-  setBackgroundImageBasedOnTime();
+  setBackgroundImage('main');
 });
