+++ src/js/utilities.mjs
@@ -32,3 +32,31 @@ export function loadDynamicContent() {
   loadFileIntoElement('/partials/header.html', 'dynamic-header');
   loadFileIntoElement('/partials/footer.html', 'dynamic-footer');
 }
+
+// Function to set a background image based on the current time
+export function setBackgroundImage(selector) {
+  // Array of background images you'd like to cycle through
+  const images = [
+      '/images/backdrop1.jpg',
+      '/images/backdrop2.jpg',
+      '/images/backdrop3.jpg',
+      '/images/backdrop4.jpg',
+  ];
+
+  const hour = new Date().getHours();
+  const mainElement = document.querySelector(selector);
+
+  let imageUrl;
+
+  if (hour >= 0 && hour < 6) {
+      imageUrl = images[0];
+  } else if (hour >= 6 && hour < 12) {
+      imageUrl = images[1];
+  } else if (hour >= 12 && hour < 18) {
+      imageUrl = images[2];
+  } else {
+      imageUrl = images[3];
+  }
+
+  mainElement.style.backgroundImage = `url('${imageUrl}')`;
+}
