+++ src/public/partials/footer.html
@@ -1,6 +1,6 @@
 
   <div class="left">
-    © Copyright 2023 All Rights Reserved
+    © Copyfakeright 2023 All Fakes Reserved
   </div>
 
   <div class="center">
