+++ src/public/partials/header.html
@@ -6,6 +6,7 @@
     </div>
     <ul class="nav-links">
           <li class="close-nav">&#10005;</li>
+          <li><p><a href="/index.html" data-replace="Home"><span>Home</span></a></p></li>
           <li><p><a href="/pages-flight/flight-list.html" data-replace="Search"><span>Search</span></a></p></li>
           <li><a href="#" data-replace="Favorites"><span>Favorites</span></a></li>
           <li><a href="/pages-others/about-us.html" data-replace="About Us"><span>About Us</span></a></li>
